backend/pyproject.toml
[tool.poetry]
name = "end-to-end-data-pipeline-backend"
version = "0.1.0"
description = "End-to-end data ingestion pipeline with Flask"
authors = ["Igwe Chibueze <scientistigwe@gmail.com>"]
readme = "README.md"
packages = [
    { include = "api" },
    { include = "core" },
    { include = "data" },
    { include = "db" },
    { include = "utils" },
    { include = "config" },
    { include = "docs" },
    { include = "infrastructure" },
    { include = "subscription" },
    { include = "tests" }
]

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
flask = "^3.0.0"
flask-cors = "^4.0.0"
flask-jwt-extended = "^4.7.1"
flask-apispec = "^0.11.4"
flask-sqlalchemy = "^3.1.1"
gunicorn = "^21.2.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.36"
sqlalchemy-utils = "^0.41.2"
alembic = "^1.14.0"

# Database Drivers
pymongo = "^4.10.1"
pymysql = "^1.1.1"
psycopg2-binary = "^2.9.10"
aiopg = "^1.4.0"
aiomysql = "^0.2.0"

# Cloud & Storage
boto3 = "^1.35.59"
aioboto3 = "^13.3.0"
botocore = "^1.35.60"

# Data Processing
pyarrow = "^18.0.0"
fastparquet = "^2024.11.0"
openpyxl = "^2.1.5"

# Security
cryptography = "^43.0.3"
pycryptodome = "^3.21.0"
python-jose = "^3.3.0"

# API & Validation
pydantic = "^2.9.2"
marshmallow = "^3.23.1"
apispec = "^6.7.1"
email-validator = "^2.2.0"

# Data Quality & Analysis
ydata-profiling = "^4.12.0"
pyspellchecker = "^0.7.2"
chardet = "^5.2.0"
jellyfish = "^1.1.0"
textstat = "^0.7.4"
sentence-transformers = "^3.3.1"

# Utilities
loguru = "^0.7.2"
tqdm = "^4.67.1"
psutil = "^6.1.0"
aiohttp = "^3.11.11"
aiofiles = "^24.1.0"
cachetools = "^5.5.0"

# Message Queues
pika = "^1.3.2"
confluent-kafka = "^2.8.0"
tabulate = "^0.9.0"
humanize = "^4.11.0"
plotly = "^5.24.1"
marshmallow-enum = "^1.5.1"
celery = "^5.4.0"
prometheus-client = "^0.21.1"
influxdb-client = "^1.48.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pre-commit = "^3.3.3"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=term-missing --asyncio-mode=auto"
asyncio_mode = "auto"